# openapi_hf_space.yaml
openapi: 3.0.3
info:
  title: Cryptocurrency Data Source API - HuggingFace Space
  version: 1.0.0
  description: |
    # مستندات API منبع داده ارز دیجیتال
    
    این API به عنوان **ارائه‌دهنده واحد** برای تمام داده‌های cryptocurrency عمل می‌کند.
    
    ## اولویت‌ها (Priority)
    1. **HF-first (HTTP)** - ابتدا از endpoints داخلی HuggingFace استفاده می‌شود
    2. **WS-only exception** - WebSocket فقط برای endpoint‌های خاص
    3. **Fallback-last** - در صورت عدم موفقیت از providers پشتیبان استفاده می‌شود
    
    ## منابع داده (Data Sources)
    - **Primary**: HuggingFace Space endpoints
    - **Fallback**: External providers (CoinGecko, Binance, WhaleAlert, etc.)
    - **Config**: `/mnt/data/api-config-complete.txt`
    
    ## Meta Fields
    تمام response‌ها شامل فیلد `meta` با اطلاعات زیر:
    - `source`: منبع داده (hf, hf-ws, یا URL provider)
    - `cache_ttl_seconds`: مدت زمان cache
    - `generated_at`: زمان تولید (ISO 8601)
    - `attempted`: لیست منابع تلاش شده (در صورت خطا)
    
  contact:
    name: Amin - Crypto Data API
    url: https://really-amin-datasourceforcryptocurrency.hf.space
  license:
    name: MIT
servers:
  - url: https://really-amin-datasourceforcryptocurrency.hf.space
    description: Production HuggingFace Space
  - url: http://localhost:7860
    description: Local Development
...
paths:
...
  /api/market:
    get:
      tags: [Market Data]
      summary: دریافت لیست بازار (Market Snapshot)
      description: |
        دریافت لیست ارزهای برتر با قیمت، حجم و تغییرات 24 ساعته.
        **Priority**: HF HTTP first
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 200
          description: تعداد آیتم‌ها
        - name: sort
          in: query
          schema:
            type: string
            enum: [price, volume, change, market_cap]
            default: market_cap
          description: نوع مرتب‌سازی
      responses:
        '200':
          description: موفقیت‌آمیز
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketResponse'
        '502':
          $ref: '#/components/responses/BadGateway'
...
  /api/sentiment/analyze:
    post:
      tags: [Sentiment]
      summary: تحلیل احساسات
      description: تحلیل sentiment یک متن
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  description: متن برای تحلیل
                  example: "Bitcoin price is rising rapidly today!"
                mode:
                  type: string
                  enum: [simple, detailed]
                  default: simple
                  description: نوع تحلیل
      responses:
        '200':
          description: موفقیت‌آمیز
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SentimentResponse'
  
  # ============================================================================
  # WHALE TRACKING ENDPOINTS
  # ============================================================================
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key برای endpoint‌های محافظت شده
  schemas:
...
    MetaInfo:
      type: object
      required:
        - source
        - generated_at
      properties:
        source:
          type: string
          description: منبع داده (hf, hf-ws, یا URL provider)
          example: hf
        cache_ttl_seconds:
          type: integer
          description: مدت زمان cache (ثانیه)
          example: 30
        generated_at:
          type: string
          format: date-time
          description: زمان تولید response (ISO 8601)
        attempted:
          type: array
          items:
            type: string
          description: لیست منابع تلاش شده (فقط در صورت خطا)
          example: ["hf", "coingecko", "binance"]
          
    # Market Data Schemas
    MarketItem:
      type: object
      required:
        - symbol
        - price
      properties:
        symbol:
          type: string
          example: BTC
        name:
          type: string
          example: Bitcoin
        price:
          type: number
          format: float
          example: 45000.50
        change_24h:
          type: number
          format: float
          example: 2.34
          description: تغییر 24 ساعته (درصد)
        volume_24h:
          type: number
          format: float
          example: 25000000000
        market_cap:
          type: number
          format: float
          example: 880000000000
        rank:
          type: integer
          example: 1
        source:
          type: string
          example: binance
    MarketResponse:
      type: object
      properties:
        last_updated:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: '#/components/schemas/MarketItem'
        meta:
          $ref: '#/components/schemas/MetaInfo'
...
    SentimentResponse:
      type: object
      properties:
        score:
          type: number
          format: float
          minimum: -1
          maximum: 1
          description: نمره sentiment (-1 تا +1)
        label:
          type: string
          enum: [positive, negative, neutral]
        details:
          type: object
          properties:
            positive:
              type: number
            negative:
              type: number
            neutral:
              type: number
        meta:
          $ref: '#/components/schemas/MetaInfo'
    
    # Whale Tracking Schemas
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: BadGateway
        message:
          type: string
          example: All providers failed
        meta:
          type: object
          properties:
            attempted:
              type: array
              items:
                type: string
              description: لیست منابع تلاش شده
              example: ["hf", "coingecko", "binance"]
            timestamp:
              type: string
              format: date-time
  responses:
    BadGateway:
      description: تمام providers شکست خوردند
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: BadGateway
            message: All upstream providers failed
            meta:
              attempted: ["hf", "coingecko", "binance"]
              timestamp: "2025-11-24T10:00:00Z"
...